Remove proposal page
