+++ nextjs_frontend/src/app/proposal/[cpfDevedor]/components/FinalProposalCard/final-proposal-card.tsx
