--- nextjs_frontend/src/app/proposal/[cpfDevedor]/components/ProposalCard/proposal-card.module.scss
@@ -1,19 +1,0 @@
-.proposalCard {
-  @apply w-full h-96 mb-10 flex flex-col items-center justify-center
-       bg-white shadow border rounded-lg border-gray-950
-         border-opacity-50 transition-transform ease-in-out duration-300;
-
-  &[data-selected="true"] {
-    @apply scale-125 -translate-y-5;
-  }
-
-  ul li {
-    @apply pl-7 pt-1 bg-[url(/icons/check_circle.svg)]
-           bg-no-repeat bg-left;
-  }
-
-  ~ button[data-selected="false"] {
-    @apply bg-white border border-neutral-400
-            text-neutral-400 font-medium;
-  }
-}
