--- nextjs_frontend/src/app/proposal/[cpfDevedor]/components/ProposalLoading/proposal-loading.module.scss
@@ -1,14 +1,0 @@
-.skeletonCard {
-  @apply w-full h-96 mb-10 flex flex-col items-center justify-center
-       bg-white shadow border rounded-lg border-gray-950
-         border-opacity-50 transition-transform ease-in-out duration-300;
-
-  ul li {
-    @apply pl-7 py-2 bg-[url(/icons/check_circle.svg)]
-           bg-no-repeat bg-left h-10 flex;
-  }
-
-  .field {
-    @apply bg-slate-200 h-full w-full flex rounded-md;
-  }
-}
