--- nextjs_frontend/src/app/proposal/[cpfDevedor]/components/ProposalPodium/proposal-podium.module.scss
@@ -1,4 +1,0 @@
-.cardWrapper {
-  @apply w-full pt-20 pb-10 px-10 flex justify-center
-         items-center gap-5 md:gap-10 xl:gap-20 max-lg:overflow-hidden;
-}
